# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LUXIMA API Documentation",
        "description": "This is the API documentation for the Luxima project. It covers all endpoints related to authentication and other functionalities.",
        "contact": {
            "email": "admin@luxima.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.luxima.id/api",
            "description": "LUXIMA API Server"
        }
    ],
    "paths": {
...
        "/bookings": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Show all Bookings",
                "description": "Retrieve a list of all Bookings.",
                "operationId": "2a8518bafccd7b95f3cb325c26f32963",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved Bookings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Booking"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized, token tidak valid"
                    }
                }
            }
        },
        "/booking/{id}": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Show detail booking by id",
                "description": "Retrieve detail data booking by id.",
                "operationId": "0c99932638468b76e93434ad208239e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Unique booking id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data Booking Detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                }
            }
        },
        "/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Show all Categories",
                "description": "Retrieve a list of all Categories.",
                "operationId": "11dc7bee7f8ad5b8d0a6a33493c5ea92",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved Categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized, token tidak valid"
                    }
                }
            }
        },
        "/category/{slug}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Show detail category by slug",
                "description": "Retrieve detail data category by slug.",
                "operationId": "22ac95e86319ac642d72e9ed0fe15a56",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Unique category slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data Category Detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
        "/locations": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Show all Locations",
                "description": "Retrieve a list of all Locations.",
                "operationId": "2be5659d02ad51e62526d661084c6d41",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved Locations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Location"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized, token tidak valid"
                    }
                }
            }
        },
        "/location/{slug}": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Show detail location by slug",
                "description": "Retrieve detail data location by slug.",
                "operationId": "c9d5f519cca98fa219c78b5e32f099dd",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Unique location slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data Vendor Detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Location"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Location not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Booking": {
                "title": "Booking",
                "required": [
                    "id",
                    "name",
                    "slug",
                    "location"
                ],
                "properties": {
                    "id": {
                        "description": " ID Booking",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "User ID Booking",
                        "type": "integer"
                    },
                    "vendor_id": {
                        "description": "Vendor ID Booking",
                        "type": "integer"
                    },
                    "service_id": {
                        "description": "Service ID Booking",
                        "type": "integer"
                    },
                    "event_id": {
                        "description": "Event ID Booking",
                        "type": "integer"
                    },
                    "location_id": {
                        "description": "Location ID Booking",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Booking name",
                        "type": "string"
                    },
                    "total_amount": {
                        "description": "Booking total amount",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Booking created at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Booking updated at",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Category": {
                "title": "Categories",
                "required": [
                    "name",
                    "slug",
                    "description",
                    "image"
                ],
                "properties": {
                    "name": {
                        "description": "Category name",
                        "type": "string"
                    },
                    "slug": {
                        "description": "Unique slug for category",
                        "type": "string"
                    },
                    "description": {
                        "description": "Category description",
                        "type": "string"
                    },
                    "image": {
                        "description": "Category image",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Location": {
                "title": "Location",
                "required": [
                    "id",
                    "name",
                    "slug",
                    "location"
                ],
                "properties": {
                    "id": {
                        "description": " ID vendor",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Location name",
                        "type": "string"
                    },
                    "slug": {
                        "description": "Unique slug for vendor",
                        "type": "string"
                    },
                    "description": {
                        "description": "Location description",
                        "type": "string"
                    },
                    "location": {
                        "description": "Location location",
                        "type": "string"
                    },
                    "category": {
                        "description": "Location category",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Location created at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Location updated at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "description": "Location deleted at",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}